% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rating.R
\name{aggregate.rating}
\alias{aggregate.rating}
\title{aggregation of annnotations data}
\usage{
\method{aggregate}{rating}(ChildRecordings, data, cut = 0.1)
}
\arguments{
\item{ChildRecordings}{: a ChildRecordings class}

\item{data}{: find.ratting.segment return or similar data.frame}

\item{cut}{: time size in second for the unit segment}
}
\value{
Soon
}
\description{
Base on the result of an find.ratting.segment return or similar data.frame
the function will extract data from annotation file in a raterData Class
}
\details{
the data will be organize un raw annotation format and a long segmented format
}
