% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Reliability analyse}
\usage{
analyse(x)
}
\arguments{
\item{raterData}{: a raterData class}
}
\value{
Soon
}
\description{
Provide a reliability analyse returnin Krippen alpha, fleiss kappa and gwet AC1
}
