% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChidRecordings.R
\name{ChildRecordings}
\alias{ChildRecordings}
\title{Building childrecordings class}
\usage{
ChildRecordings(path)
}
\arguments{
\item{path}{path to the childrecoding project folder}
}
\value{
A ChildRecordings class
}
\description{
Main class of the package. Help to check data integrity and provide data
organization. Only provide the path folder of the childrecordings project.
Provid some integrity check such as missing files or missing meta
}
