% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertor_long_cut.R
\name{convertor_long_cut}
\alias{convertor_long_cut}
\title{Converte annotation format to long format}
\usage{
convertor_long_cut(data, onset, offset, cut = 0.1)
}
\arguments{
\item{data}{: a data.frame annotation}

\item{onset}{offset : time of the data where the conversion should start and stop}

\item{cut}{: lenght of the windoaws in second}
}
\description{
Convert annotation data.frame to a long format sequence when the length of the
 time windows is define by the cut
}
