% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertor_long_cut_loop}
\alias{convertor_long_cut_loop}
\title{Rcpp function to replace the for-loop inside the 'convertor_long_cut()' function}
\usage{
convertor_long_cut_loop(
  time_seq,
  segment_onset_vec,
  segment_offset_vec,
  speaker_type_vec
)
}
\arguments{
\item{time_seq}{a numeric vector of the time sequence}

\item{segment_onset_vec}{a numeric vector of onset}

\item{segment_offset_vec}{a numeric vector of offset}

\item{speaker_type_vec}{a character vector of speaker type}
}
\value{
a named list of length 2
}
\description{
Rcpp function to replace the for-loop inside the 'convertor_long_cut()' function
}
