% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDataCR.R
\name{extractDataCR}
\alias{extractDataCR}
\title{file importation from annotator}
\usage{
extractDataCR(ChildRecordings, set.type)
}
\arguments{
\item{ChildRecordings}{: a ChildRecordings class}

\item{set.type}{: a string containing th name of the annotator present in the "set" column meta}
}
\value{
: a data.frame
}
\description{
Import all the file from a specific annotators
}
