% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_opener.R
\name{file.opener}
\alias{file.opener}
\title{Opening a data file in ChildRecordingsData class}
\usage{
file.opener(meta_row, ChildRecordings, use_data_table = FALSE, threads = 1)
}
\arguments{
\item{meta_row}{: a row from the meta containing all the info of the cvs file}

\item{ChildRecordings}{: a ChildRecordingsData class}

\item{use_data_table}{: use the data.table package to read the .csv annotation data (depending on the operating system and the number of threads used it can be 3 to 5 times faster than 'read.csv')}

\item{threads}{the number of threads to run in parallel}
}
\value{
A data.frame
}
\description{
Open a csv file from a ChildRecordingsData class.
The function provides transformation of data such as creating a true segment_onset
and segment_offset based on meta data. It will also provide POSIXct time info if
start_time of the recording is provided in the meta
}
\examples{

\dontrun{

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/vandam-daylong-demo"
CR = ChildRecordings(path)
raw_file <- file.opener(CR$all.meta[1,],CR, use_data_table = TRUE,
                        threads = parallel::detectCores())
head(raw_file)

}
}
