% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_openner.R
\name{file.openner}
\alias{file.openner}
\title{Openning a data file in childrecording class}
\usage{
file.openner(meta_row, ChildRecordings)
}
\arguments{
\item{meta_row}{: a row from the meta containing all the info of the cvs file}

\item{ChildRecordings}{: a ChildRecordings class}
}
\value{
A data.frame
}
\description{
Open a csv file from a ChildRecordings class.
The function provide transformation on data such as create a true segment_onset
and segment_offset based on meta data. Will also provide POSIXct time info if
strat_time of the recording is provided in the meta
}
\examples{
library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/namibia-data/"
CR = ChildRecordings(path)
raw_file <- file.openner(CR$all.meta[1,],CR)
head(raw_file)

}
