% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ratting_segment.R
\name{find.ratting.segment}
\alias{find.ratting.segment}
\title{Search function on Childrecordings class}
\usage{
find.ratting.segment(
  ChildRecordings,
  filename,
  annotators = NULL,
  range_from = NULL,
  range_to = NULL
)
}
\arguments{
\item{ChildRecordings}{: a ChildRecordings class}

\item{filename}{: a wav file name to look}

\item{annotators}{: an optional argument providing le annotators to look at}

\item{range_from}{and range_to : an optional time windows for the search}
}
\value{
A data.frame containing commun annoation segment for each annotator
}
\description{
Search function in Childrecordings class
This search function will help you to provide file name and time windows
for common coding in raters
}
