% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{get.reliability}
\alias{get.reliability}
\title{Reliability analyse}
\usage{
get.reliability(raterData, summary = TRUE)
}
\arguments{
\item{raterData}{: a raterData class}
}
\value{
Soon
}
\description{
Provide a reliability analyse returnin Krippen alpha, fleiss kappa and gwet AC1
}
\examples{
library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

# if no time windows is specified, this function will only return at table for all the know raters
# All the rater need to ratter any segment find
search = find.rating.segment(CR, "aiku/namibie_aiku_20160715_1.wav")
rez = aggregate.rating(search, CR, cut=0.100, verbose=T)
get.reliability(rez)

}
