% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LENA_overlap_loop}
\alias{LENA_overlap_loop}
\title{Rcpp function to replace the for-loop inside the 'LENA.overlap()' function}
\usage{
LENA_overlap_loop(segment_onset_vec, segment_offset_vec, speaker_type_vec)
}
\arguments{
\item{segment_onset_vec}{a numeric vector of onset}

\item{segment_offset_vec}{a numeric vector of offset}

\item{speaker_type_vec}{a character vector of speaker type}
}
\value{
a named list of length 2
}
\description{
Rcpp function to replace the for-loop inside the 'LENA.overlap()' function
}
