% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterComparaison.R
\name{raterComparaison}
\alias{raterComparaison}
\title{Compare Annotator reliability}
\usage{
raterComparaison(raterData)
}
\arguments{
\item{raterData}{: a raterData class}
}
\value{
Soon
}
\description{
Provide indication of quality for each annotator.
Following classical test theory, each annotator is removed from the data and a reliability indicator is provided for the annotation as a whole.
Removing a good annotator should decrease the indicator.
In signal detection theory a mean of precision, recall and F-score are provided by rater.
}
\examples{
library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

# if no time windows is specified, this function will only return at table for all the know raters
# All the rater need to ratter any segment find
search = find.rating.segment(CR, "aiku/namibie_aiku_20160715_1.wav")
rez = aggregate.rating(search, CR, cut=0.100, verbose=T)

comparaison = raterComparaison(rez)
plot(comparaison)
}
