% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDataCR.R
\name{summary.extractDataCR}
\alias{summary.extractDataCR}
\title{statistics from imported annotator}
\usage{
\method{summary}{extractDataCR}(extractDataCR)
}
\arguments{
\item{extractDataCR}{: a extractDataCR class}
}
\value{
A list with summary data and indicators,  ICC and regression for child speaker.
}
\description{
statistics from imported annotator
}
\details{
The summary function for extractDataCR provides the computational methods for the classical indicators for speaker_type.
These indicators are then used to evaluate the interclass correlation in a mixed lme4 model.
The ICC should indicate how similar the units in the same group are.
The indicators will also be analyzed in a mixed LME4 model to provide insight into the impact of age on these indicators.
In this model, the indicator and age in days are scaled to provide a comparison between the indicators.
}
\examples{
library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)
rez = extractDataCR( "textgrid_m1", CR, verbose = F )
head(rez)
# With LENA overlap method
rez = extractDataCR("textgrid_m1", CR, LENA.OL = T, verbose = F)
summary(rez)

}
