% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_time_seg_finder.R
\name{true_time_seg_finder}
\alias{true_time_seg_finder}
\title{true time segment finder}
\usage{
true_time_seg_finder(range_from, range_to, all.meta.table)
}
\description{
true time segment finder
}
\keyword{internal}
