% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rating.R
\name{aggregate.rating}
\alias{aggregate.rating}
\title{aggregation of annnotations data}
\usage{
\method{aggregate}{rating}(data, ChildRecordings, cut = 0.1, verbose = T)
}
\arguments{
\item{data}{: find.ratting.segment return or similar data.frame}

\item{ChildRecordings}{: a ChildRecordings class}

\item{cut}{: time size in second for the unit segment}
}
\value{
Soon
}
\description{
Base on the result of an find.ratting.segment return or similar data.frame
the function will extract data from annotation file in a raterData Class
}
\details{
the data will be organize un raw annotation format and a long segmented format
}
\examples{
library(ChildRecordsR)
path = "A_childrecord_data_path"
CR = ChildRecordings(path)

# finding segments on wav file for designated rater
raters <- c("Coder1","Coder2","Coder3")
search <- find.ratting.segment(CR,"Wav_file_name", raters, range_from = t1, range_to = t2)
ratting1  = aggregate.rating(search, CR, 0.1)

# try to analyse a larger number of file
wave_file <- unique(CR$all.meta$filename) # get all the wav files
ratters <- c("Coder1","Coder2","Coder3") # Define raters you are interested in

# bind all the results
search2 <- data.frame()
for (file in wave_file[1:10]){
  print(file)
  search2 <- rbind(search2, find.ratting.segment(CR, file, ratters)) # could take some time
}
ratting2  = aggregate.rating(search2, CR, 0.1)


}
