% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDataCR.R
\name{extractDataCR}
\alias{extractDataCR}
\title{file importation from annotator}
\usage{
extractDataCR(set.type, ChildRecordings, LENA.OL = F, verbose = T)
}
\arguments{
\item{set.type}{: a string containing th name of the annotator present in the "set" column meta}

\item{ChildRecordings}{: a ChildRecordings class}

\item{LENA.OL}{: add the LENA overlap method. The overlap method will correct onset and offset when speech overlap by removing part that overlap.}
}
\value{
A class extractDataCR with a data.frame with all the aggregated data
}
\description{
Import all the files from a specific annotator
}
\examples{
library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/vandam-daylong-demo"
CR = ChildRecordings(path)
rez = extractDataCR( "vtc", CR, verbose = F )
head(rez$data)
# With LENA overlap method
rez = extractDataCR("vtc", CR, LENA.OL = T, verbose = F)


}
