% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_raterData.R
\name{get.classification}
\alias{get.classification}
\title{Signal detection theory to compare two annotators}
\usage{
get.classification(raterData, raters, plot = TRUE, summary = TRUE)
}
\arguments{
\item{raterData}{: a raterData class}

\item{raters}{: a vector containing two string with the names of the two raters}

\item{plot}{a boolean. If TRUE then the plot will be returned}

\item{summary}{a boolean. If TRUE then the summary statistics will be printed out in the console}
}
\value{
a Class to print a graph and return a list containing a confusion matrix class from caret package and a macro SDT indicator
}
\description{
Provide a SDT indicator to compare two rater and graph
}
