% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_annotation.R
\name{New.annotations}
\alias{New.annotations}
\title{Create new dummy annotation}
\usage{
New.annotations(
  row.meta,
  time.sd = NULL,
  change.cat.prob = NULL,
  new.name = NULL,
  ChildRecordings
)
}
\arguments{
\item{row.meta}{: a row meta from ChildRecordings$all.meta to transform into dummy annotation}

\item{time.sd}{: a time deviation in millisecond to apply to the segments on using a normal law}

\item{change.cat.prob}{: The probability of a change in speaker_type}

\item{new.name}{: a string containing a new annotator name}

\item{ChildRecordings}{: a ChildRecordings class}
}
\description{
Use a real annotation to create a dummy annotation with degraded ratings and
directly add info to the ChildRecordings class
}
\examples{

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/vandam-daylong-demo"
CR = ChildRecordings(path)

time.sd = 300
change.cat.prob = 0.05
row.meta = CR$all.meta[2,]
new.name = "vtc_mod"
New.annotations(row.meta, time.sd, change.cat.prob, new.name, CR)
}
