% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterComparaison.R
\name{raterComparaison}
\alias{raterComparaison}
\title{Compare Annotator reliability}
\usage{
raterComparaison(raterData)
}
\arguments{
\item{raterData}{: a raterData class}
}
\description{
Compare a rater from a raterData class and provide indication of quality for each annotator.
In classical test theory tested annotator is retracted from the data and reliabily indicator are provided.
A good annotator should should decrase the indicator after retraction.
In signal detection theory a mean of precision, recal and F-score are provided by rater.
}
