% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_raterData.R
\name{SDT.raterData}
\alias{SDT.raterData}
\title{Signal detection theory to compare two annotators}
\usage{
SDT.raterData(raterData, raters)
}
\arguments{
\item{raterData}{: a raterData class}

\item{raters}{: a vector containning two string witht the names of the two raters}
}
\value{
Soon
}
\description{
Provide a SDT indicator to compare two rater
}
