% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_raterData.R
\name{SDT.raterData}
\alias{SDT.raterData}
\title{Signal detection theory to compare two annotators}
\usage{
SDT.raterData(raterData, raters, plot = TRUE, summary = TRUE)
}
\arguments{
\item{raterData}{: a raterData class}

\item{raters}{: a vector containing two string with the names of the two raters}
}
\value{
a Class,  print a graph and return a list containing a confusion matrix class from caret package and a macro SDT indicator
}
\description{
Provide a SDT indicator to compare two rater and graph
}
\examples{
library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

# finding segments on wav file for designated rater
raters <- c("textgrid_ak","textgrid_mm","textgrid_m1")
search <- find.rating.segment(CR,"aiku/namibie_aiku_20160715_1.wav",range_from = 27180, range_to = 27240)
ratting1  = aggregate.rating(search, CR, 0.1)
comp1 = raterComparaison(ratting1)
SDT.raterData(comp1,c("textgrid_ak","textgrid_mm"))


}
